Move Rocket integration into askama_rocket crate
