+++ askama/Cargo.toml
@@ -25,7 +25,7 @@ serde-json = ["askama_shared/json"]
 serde-yaml = ["askama_shared/yaml"]
 num-traits = ["askama_shared/num-traits"]
 with-iron = ["iron", "askama_derive/iron"]
-with-rocket = ["rocket", "askama_derive/rocket"]
+with-rocket = ["askama_derive/rocket"]
 with-actix-web = ["askama_derive/actix-web"]
 with-gotham = ["gotham", "askama_derive/gotham", "hyper", "mime", "mime_guess"]
 
@@ -34,7 +34,6 @@ askama_derive = { version = "0.9.0", path = "../askama_derive" }
 askama_escape = { version = "0.3.0", path = "../askama_escape" }
 askama_shared = { version = "0.9.0", path = "../askama_shared", default-features = false }
 iron = { version = ">= 0.5, < 0.7", optional = true }
-rocket = { version = "0.4", default-features = false, optional = true }
 mime = { version = "0.3", optional = true }
 mime_guess = { version = "2.0.0-alpha", optional = true }
 gotham = { version = "0.3", default-features = false, optional = true }
