+++ askama_rocket/tests/basic.rs
@@ -1,4 +1,3 @@
-#![cfg(feature = "with-rocket")]
 #![feature(proc_macro_hygiene, decl_macro)]
 
 #[macro_use]
