+++ testing/Cargo.toml
@@ -9,14 +9,12 @@ edition = "2018"
 default = []
 full = ["with-iron", "serde-json", "with-gotham"]
 serde-json = ["serde_json", "askama/serde-json"]
-with-rocket = ["rocket", "askama/with-rocket"]
 with-iron = ["iron", "askama/with-iron"]
 with-gotham = ["gotham", "askama/with-gotham", "mime", "hyper"]
 
 [dependencies]
 askama = { path = "../askama", version = "*" }
 iron = { version = "0.6", optional = true }
-rocket = { version = "0.4", default-features = false, optional = true }
 serde_json = { version = "1.0", optional = true }
 gotham = { version = "0.3", default-features = false, optional = true }
 mime = { version = "0.3", optional = true }
